realtime_interval -> block_interval / 2
The realtime_interval should be half of the block_interval to ensure
that we don't get stuck in a state where we just miss a new block
because using precisely the block rate doesn't account for processing
time.  The exact value of half the block_interval was chosen for the
Nyquist rules.