Subscribe to newHeads for realtime indexer
Separate Catchup and Realtime indexer's fetching, so that Catchup
continues to use the old polling, `EthereumJSONRPC.json_rpc` while Realtime
subscribes with `EthereumJSONRPC.subscribe`.

Because Realtime gets notified of new blocks through the subscription,
it no longer has a Sequence, so Sequence has been removed from
`%BlockFetcher{}`.  There is also no block concurrency or block batch
size for Realtime, so that was removed from `%BlockFetcher{}` too and
moved to `%BlockFetcher.Catchup{}`.

Because Realtime requires a websocket to be connected, there is now an
`Indexer.BlockFetcher.Realtime.Supervisor` that supervises the
`Indexer.BlockFetcher.Realtime.WebSocket` and
`Indexer.BlockFetcher.Realtime`.  `Indexer.BlockFetcher.Realtime`
because it is waiting on notifications from is subscription becomes a
`GenServer` insteado of a `Task`, so it can be supervised normally.

BlockFetcher.Supervisor becomes a normal `Supervisor` of
`Catchup.Supervisor` and `Realtime.Supervisor`.

As Realtime is using `EthereumJSONRPC.subscribe`,
`subscribe_named_arguments` has been added as a sibling of
`json_rpc_named_arguments` in the `:indexer` `config`s.