Add generic Ordered Cache behaviour and implementation
Problem: multiple caches exist to keep a page of the most recent `n` elements.
More are to be implemented, so it would be good to abstract their implementation.

Solution: add a behaviour and a macro (providing an implementation) to create this type of caches.
Redefine the existing ones using the new macro.